fix signed/unsigned comparison warning

+++ mvptable.cpp
@@ -577,7 +577,8 @@ MVPTable::blob_type MVPTable::find_vantage_point(size_t sample_size)
 		// since we don't have any vantage points yet, we will rank
 		// the points by the sum of the distances to all of the other points
 		// and pick the greatest
-		if (num_points <= sample_size) {
+		//the static_cast to unsigned is OK because a previous case confirmed num_points > 0
+		if (static_cast<size_t>(num_points) <= sample_size) {
 			//we have few points, so do the pairwise distance between all
 			return cache.exec_getBlob(
 				"SELECT value FROM ("
